Move specs used in cucumber-jvm tests
to where they are used
