+++ console-runner-features/src/test/java/info/javaspec/example/main/OneFailsSpecs.java
@@ -1,11 +1,11 @@
-package info.javaspec.console;
+package info.javaspec.example.main;
 
 import static info.javaspec.lang.lambda.FunctionalDsl.describe;
 import static info.javaspec.lang.lambda.FunctionalDsl.it;
 
-public final class MainStepsOneFails {
+public final class OneFailsSpecs {
   {
-    describe("MainStepsOneFails", () -> {
+    describe("OneFails", () -> {
       it("fails", () -> { throw new AssertionError("bang!"); });
     });
   }
