+++ console-runner-features/src/test/java/info/javaspec/example/main/OneOfEachSpecs.java
@@ -1,4 +1,4 @@
-package info.javaspec.console;
+package info.javaspec.example.main;
 
 import java.util.LinkedList;
 import java.util.List;
@@ -8,7 +8,7 @@ import static info.javaspec.lang.lambda.FunctionalDsl.it;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.contains;
 
-public final class MainStepsOneOfEach {
+public final class OneOfEachSpecs {
   private static final List<String> _specsRan = new LinkedList<>();
 
   public static void reset() {
@@ -20,7 +20,7 @@ public final class MainStepsOneOfEach {
   }
 
   {
-    describe("MainStepsOneOfEach", () -> {
+    describe("OneOfEach", () -> {
       it("passes", () -> _specsRan.add("passes"));
       it("fails", () -> {
         _specsRan.add("fails");
