+++ console-runner-features/src/test/java/info/javaspec/example/main/OnePassesSpecs.java
@@ -1,11 +1,11 @@
-package info.javaspec.console;
+package info.javaspec.example.main;
 
 import static info.javaspec.lang.lambda.FunctionalDsl.describe;
 import static info.javaspec.lang.lambda.FunctionalDsl.it;
 
-public final class MainStepsOnePasses {
+public final class OnePassesSpecs {
   {
-    describe("MainStepsOnePasses", () -> {
+    describe("OnePasses", () -> {
       it("passes", () -> {});
     });
   }
